tasks/generate_assessment_chart_configs/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*